created unified error page
