+++ package.json
@@ -26,5 +26,8 @@
     "@typescript-eslint/parser": "^6.21.0",
     "husky": "^9.1.7",
     "lerna": "^5.4.3"
+  },
+  "dependencies": {
+    "client-server-template-with-vite": "file:"
   }
 }
