+++ packages/client/src/pages/error/ui/index.ts
@@ -1,2 +1,1 @@
-export { NotFoundPage } from './NotFoundPage';
-export { ServerErrorPage } from './ServerErrorPage';
+export { ErrorPage } from './ErrorPage';
